internal: Add flake.nix
This commit adds a flake.nix with a dev-shell and formatting configured.

I decided that it would be a good idea to enable formatting for our nix
files immediately, so that we don't need to worry about reformatting
existing Nix code down the line. I didn't add any other formatters for
now however.