test: Add Tree-sitter checks to flake.nix
Added two new checks to the Nix flake, checking that
`tree-sitter generate` has been ran, and that the tests in the 'test/'
directory pass.

Also note that it seems like regenerating the parser changed a some of
the files under `src/`. I guess those were due to changes in 'parser.js'
that hadn't been regenerated in an earlier commit and/or that the version of the
`tree-sitter` CLI I used did things slightly differently.