feat: Add more supertypes
Adds the following new hidden nodes marked as supertypes:
- `_callable`: Supertype of `function`, `closure`,
  `function_definition`, `macro`, and `macro_definition`.
- `_matcher`: Supertype of `match`, `recv`, and `recv_timeout`.
- `_loop`: Supertype of `loop`, `loopfor`, `loopwhile`, `looprange`,
  `loopforeach`, and `loopwhile_thread`.

Since the nodes in the respective categories have similar children with
a consumer (say, a formatter) may find it more useful to match against a
common supertype, instead of having to write separate query patterns for
each subtype node. Additionally, the fields which are shared amongst the
node types were originally designed so that the matching fields always
contain nodes with "similar semantics". So having the super types means
you can more easily exploit these shared fields.

I also renamed the 'timeout' field of `recv_timeout` to 'value' to match
the field of the same name in `match`. They both contain the first value
which is an evaluated expression in `recv-to` and `match` forms. So it's
likely that you'd want to treat them the same.