feat: Add support for hex literals
Adds support for hexadecimal integer literals like "0x48f1". Also
changes the number node so that any minus signs are tokenized separately
and placed in a new "operator" field.

This required changing the implementation of the `symbol` node slightly
by making a literal "-" token a possible symbol, instead of that said
string being matched by one of the regex patterns.
I couldn't get it to parse correctly without this change, but I'm not
sure there isn't be any other way of solving this.